fixed issue with player sticking to walls
add force does not cooperate with physics materials, which means that as long as there is a force being added, the player won't "slide". to fix this, the player now has a grounded check, and when it is false they wont be able to control themselves (which means that add force won't stop the player from sliding though the gaps in the map). a physics material has also been added, meaning that we can change the amount that a player slides when they aren't inputing